Support indexed columns in PK/UK constraints
Summary:
This will allow PK and UK constraints to contain expressions and sort order specifications rather than just column names.

Indexed_columns are now the norm for any index, and hence our grammar matches the SQLite grammar. Most of the foundation was already done so this is just a little wiring.

I did not rename the tree nodes in this diff because that would make the .ref file change huge; to keep it reviewable, I will do a simple rename in a later diff that is only a rename.

Reviewed By: RaoulFoaleng

Differential Revision: D28016552

fbshipit-source-id: a358e9e8e156c5afa572408f494ed57e6e5f8a99