Demand an exact type match on the columns added in an extension proc
Summary:
I fixed a lingering bug in the verifier so that it is clear which of the type is the expected type and which is the actual type
I fixed the call sites accordingly.
The expected type is no longer marked with an error if there is no match, only the actual type hence the type verifier is no longer destructive.

Reviewed By: RaoulFoaleng, winniequinn

Differential Revision: D28805124

fbshipit-source-id: 2140a9eac08ad8dccdbea455a30fc08b2e73dad6