Emit correct nullability annotations for OUT NOT NULL
Summary:
D27721952 introduced the use of contracts to enforce nullability. In all cases *except for one*, the contracts emitted matched the nullability annotations of the corresponding C function parameters.

This diff eliminates that one exception. In the case of a `T ... OUT NOT NULL` parameter, where `T` is one of `TEXT`, `BLOB`, or `OBJECT`, we now annotate such that the pointer being pointed to is nullable. For example, in the case of `TEXT t OUT NOT NULL`, we now emit `cql_string_ref _Nullable *_Nonnull t`. This makes sense because `*t` will always be `NULL` before the C procedure is called because the procedure is meant to initialize `*t`.

A little experimentation shows that this change does not seem to affect warnings issued by Clang at all. For example, you can freely pass a value of type `cql_string_ref _Nullable *_Nonnull t` even if a `cql_string_ref _Nonnull *_Nonnull t` is expected. Likewise, you can, and *should*, declare a `cql_string_ref _Nonnull *_Nonnull t` local to be initialized by calling a procedure with a `TEXT t OUT NOT NULL` parameter even though that parameter will now have the type `cql_string_ref _Nullable *_Nonnull t`.

Reviewed By: ricomariani

Differential Revision: D28520018

fbshipit-source-id: 46e714cf0847c2df9c3ead552c858b0cb5eaa65f