Codegen contracts to enforce nullability
Summary: When calling generated functions from C, it is easy to inadvertently pass NULL when a non-NULL pointer is required. The same goes for cases in which one must pass a non-NULL pointer that itself points to a non-NULL pointer (as in the case of an `INOUT ... TEXT NOT NULL` argument). By inserting contracts into the generated code, we can catch these errors earlier than they'd otherwise be caught and clearly indicate the problem to the programmer.

Reviewed By: ricomariani

Differential Revision: D27721952

fbshipit-source-id: 454f3686ee3bf34ea5a4456453840e3ea51b944b