Compile the getters version of the codegen also in codegen test (#60)
Summary:
This just verifies that when we build the getters code-gen test results that they actually compile.  This helps us to catch unused variables and whatnot before they cause warnings for users.  The actual validation of output is done in the verification step but running clang often finds issues.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/60

Test Plan: this is a test only diff, these are the tests

Reviewed By: wanaya

Differential Revision: D29068486

Pulled By: ricomariani

fbshipit-source-id: 0a5dbfa0bc8a5d59843946cefadd97971f560105