allow recreate tables to be deleted "in the past"
Summary:
The idea here is that you might delete the table in say version 4.  Then some other developer comes along and adds version 5
before you can checkin.  You merge with this but now your change is invalid.

It doesn't really matter what version the tombstone for a recreate table appears in, as long as there is one.  So we can relax the rules here so that you get one less kind of merge conflict that you can't deal with.

Error messages adjusted and tests added

Reviewed By: winniequinn

Differential Revision: D29076078

fbshipit-source-id: 62c5ef0113f2392efb249108e96deffdcdb1f30a