add cql:generate_copy as attribute
Summary:
This stack try to replace the use of `--generate_copy` as an optional command line parameter.

Most of our buck tooling have by default this parameter enabled, this will help to enable copy capabilities only in the queries we need it.

**What to review?**

In this diff we are adding a new attribute to be parsed `cql:generate_copy`.

We are adding this check together with the command `--generate_copy` to be removed in a latter diff.

We also adding this check to the objc and android codegen, those codegen were always generating a copy function even without `--generate_copy` in place.

Reviewed By: ricomariani

Differential Revision: D29086027

fbshipit-source-id: 44478cc7625171cc2a0e0884d22c07af9904a8d2