add :generate_copy to sproc
Summary:
This stack try to replace the use of --generate_copy as an optional command line parameter.
Most of our buck tooling have by default this parameter enabled, this will help to enable copy capabilities only in the queries we need it.

**What to review?**
In this diff we identify which sproc are the ones that are being copied in the code base and we add the new attribute.

Reviewed By: ricomariani

Differential Revision: D29087790

fbshipit-source-id: 10e08b2731f20c9ff62d547a1cac741328505496