remove --generate_copy param
Summary:
This stack try to replace the use of --generate_copy as an optional command line parameter.
Most of our buck tooling have by default this parameter enabled, this will help to enable copy capabilities only in the queries we need it.

**What to review?**
In this diff we remove completely the use of generate_copy from the command line tooling

Reviewed By: ricomariani

Differential Revision: D29103549

fbshipit-source-id: 99cbc126589de11ebfa762565fe87c397b4743d2