Fixed a lingering case where fragment validation was done with a contract
Summary:
The case where the base fragment has more than one CTE should have been caught with an error check like the others.

About a year ago I went through these and corrected all the cases where a Contract was used instead of an error check
because it's not a Contract.  This one was missed.

Reviewed By: winniequinn

Differential Revision: D29207484

fbshipit-source-id: 57e2891412ff5d5070ff8234d1a0652c51941696