emit the prototype for the assembly fetch results function with the base compilation unit
Summary:
The idea here is that even if you are say "core" code that doesn't know what the extensions will be, you just define the base,
you already know the signature of the assembled function so you can call it.  It can be assembled elsewhere later.
The signature is fixed by the base fragment so it's always safe to do this.

Reviewed By: winniequinn

Differential Revision: D29183368

fbshipit-source-id: 8d056cbb2a991e5d04949652d0b0602b01bd8f3c