select statements that are not in a procedure are now ignored as we do for DDL
Summary:
A so called "loose" select is useless in the global scope because nothing can capture its result statement.

The auto-generated global proc has a simple signature that does not include a statement out arg.  If you want to select
stuff in the global context you need to call procedures or use cursors which again makes sense because how could
anyone possibly consume the global proc select result with no type support.

So we ignore it like we do with DDL.

Note: we don't disallow it entirely, there are hundreds of loose selects in the semantic tests, we just use them to do validation.

So we don't make it an error.

We know there aren't actually any of these out there because compilation would fail if there was.  But now we won't generate bogus code
that can't possibly compile.

Reviewed By: winniequinn

Differential Revision: D29243925

fbshipit-source-id: 4d9a905abeee0d8097c583c13dcb23cfb8205ce2