extension fragment row count helper now uses static inline like the others
Summary:
Having an external for this case created unnecessary binary bloat and bogus uncovered code.  So we treat the row count getter the same as the other getters.

{F627215510}

Weird embedded codelines shown in the picture.  This was due to the count getter being emitted standalone with no other file context and hence no meaningful #line directive.

Reviewed By: wanaya

Differential Revision: D29316385

fbshipit-source-id: 5c63fb0ce7b6560d2e93fec3938270f0ee8a32b6