add reference test for codegen of base fragment
Summary:
Add reference tests for codegen of just a base fragment (without any extensions or assembly).  This gives a comparison between what happens with the extension and/or assembly visible or just the base fragment.

NOTE: There is no java codegen for the base fragment, so we have no test for that (no output file).  The Obj-C version is generated, but effectively empty.

Reviewed By: ricomariani

Differential Revision: D29341090

fbshipit-source-id: ab6f1d4a7096f6e6cef29f9d18052e0913579dd0