Add occurrence typing for nullability
Summary:
This implements occurrence typing, also known as flow typing, specifically for nullability.

As in other languages (e.g., Racket, where occurrence typing originated, and more recently Kotlin, TypeScript, et cetera), it is possible to use a conditional to refine a type. For example:

```
DECLARE PROC foo()
BEGIN
  DECLARE x INT;
  CALL some_proc_with_an_out_arg(x);
  IF x IS NOT NULL THEN
    -- `x` is NOT NULL in this context
  END IF;
  -- `x` reverts to being nullable here
END;
```

For now, for a name `n` to be improved in some conditional, it must be along the outermost spine of `AND`-separated expressions, and it must be exactly of the form `n IS NOT NULL`. For example, this will work for both `n1` and `n2`:

```
IF n1 IS NOT NULL and (x > 20 AND n2 IS NOT NULL) THEN … END IF;
```

However, neither will be improved here:

```
IF n1 IS NOT NULL OR (n2 IS NOT NULL AND n1 IS NOT NULL) THEN … END IF;
```

An improvement persists up until a `SET`, `FETCH`, or use in an `OUT` position that involves the improved variable. For example:

```
DECLARE x INT;
IF x IS NOT NULL THEN
   -- `x` is NOT NULL here
   SET x := 42;
   -- `x` is nullable here
  IF x IS NOT NULL THEN
    -- `x` is once again NOT NULL
    FETCH c INTO x;
    -- `x` is again nullable
  END IF;
  IF x IS NOT NULL THEN
    -- `x` is NOT NULL here
    CALL some_proc_with_an_out_arg(x);
    -- `x` is nullable
  END IF;
END IF;
```

**NOTE:** The above analysis with respect to `SET` and `FETCH` is overly conservative and will be enhanced in the future.

**NOTE:** Improvements are not applied if the name mentioned is in the global scope. This may also be enhanced in the future.

The above examples all show improvements of a variable, but improvements also work for cursor fields:

```
IF c.f IS NOT NULL THEN
  -- `c.f` is NOT NULL here
  FETCH c;
  -- `c.f` is again nullable
END IF;
```

In addition to `IF … THEN … END IF`, improvements are also supported in queries via `WHERE` clauses. For example:

```
CREATE TABLE t (x int, y int);
DECLARE c CURSOR FOR
  SELECT x + x as a, y + y as b
  FROM t
  WHERE x IS NOT NULL
    AND b IS NOT NULL;
FETCH c;
-- `c.a` is NOT NULL here because `x` was NOT NULL, thus so too was `x + x`
-- `c.b` is NOT NULL here because the alias `b` was NOT NULL
```

`JOIN` supports improvements for result columns as specified in a `WHERE` clause. Improvements in `ON` are not yet supported.

For any of this to work, one must first do the following:

```
ENFORCE_STRICT NOT NULL AFTER CHECK;
```

This is, most likely, a temporary situation that makes it possible to roll this out to users gradually. In the future, this functionality will probably be the default.

In terms of the implementation, improvements are implemented by rewriting the AST such that calls to `cql_inferred_notnull` unfailingly coerce a nullable input to a nonnull result, changing representations as appropriate:

```
$ cat a.sql
ENFORCE_STRICT NOT NULL AFTER CHECK;

CREATE PROC foo()
BEGIN
  DECLARE a INT;
  IF a IS NOT NULL THEN
    LET b := a;
    SELECT a;
  END IF;
END;

$ cql --in a.sql --sem
ENFORCE_STRICT NOT NULL AFTER CHECK;

CREATE PROC foo ()
BEGIN
  DECLARE a INTEGER;
  IF a IS NOT NULL THEN
    LET b := cql_inferred_notnull(a);
    SELECT cql_inferred_notnull(a);
  END IF;
END;
```

For C, the code for the conversion is generated in `cg_func_cql_inferred_notnull`. For SQLite, `cql_inferred_notnull` is no-op.

Much of the work during semantic analysis happens via `find_mutable_type`. Through the ability to look up a type in the environment and set/unset `SEM_TYPE_INFERRED_NOTNULL`, semantic analysis keeps track of the nullability of a binding throughout the flow of the program.

Reviewed By: ricomariani

Differential Revision: D29316573

fbshipit-source-id: b31408bb39601b63d48522e693c3ba470858bcef