Add hover support for cql file in VSCode
Summary:
Add hover support for CQL file in VSCode. This diff provide a base implementation of the hover support for CQL file.

As an initial diff for hover feature. It implement a generic hover info for all token of CQL language except `table name` reference where we show the schema info of the table.

**TODO:** There are still a lot of work left to cover all tokens: column name, variable names, cursor name, proc name, func name etc.

Reviewed By: winniequinn

Differential Revision: D28077044

fbshipit-source-id: b45cbbfe686fc8c5e45fc51e3d88a6c7b56ce3ca