Clean up identifier resolution
Summary:
`sem_resolve_id` and associated functions have been restructured into a series of ordered resolvers. Each resolver now takes four arguments:

- an optional AST for recording semantic information and reporting/recording errors;
- a mandatory name;
- an optional scope;
- and a mandatory pointer to a `sem_t *` that will be set to a type in the current "environment" (should one exist for the name and scope provided).

Each resolver makes its own decision as to whether or not it is the correct resolver for the given name and scope. If a resolver decides it is not the correct one to be used, it returns `SEM_RESOLVE_CONTINUE` and another resolver is tried. If a resolver decides it is the correct resolver, `SEM_RESOLVE_STOP` is returned instead—regardless of whether or not semantic analysis was successful. If an error occurs during resolution, resolvers will call `report_resolve_error` and `record_resolve_error`, both of which are no-ops if no AST was provided to the resolver in question.

The resolvers are tried in order in a way that exactly preserves the behavior of the previous implementation, including its quirks (e.g., enum cases always take silent precedence over cursor fields).

Reviewed By: ricomariani

Differential Revision: D29416930

fbshipit-source-id: 3a9eb1cdb26ff3d203671133c758192086b4604c