Prevent duplicate not null rewrites
Summary:
It was the case that if the same expression containing a nonnull-improved identifier were semantically analyzed twice, as can occur in a CTE, it would be rewritten twice. This resulted in an error as `cql_inferred_notnull` requires a nullable argument and it won't get one when applied to the result of another `cql_inferred_notnull` call. The same problem was occurring when analyzing the product of a previous CQL run that produced calls to `cql_inferred_nonnull`.

The fix here is simply to disable rewriting when analyzing a call to `cql_inferred_nonnull`.

----

This example, taken from the test added and run without the fix from this diff in place, demonstrates the problem:

```
$ cat a.sql
enforce_strict not null after check;

create proc improvements_work_within_ctes()
begin
  declare a int;
  if a is not null then
    with recursive foo(b) as (select a)
    select b from foo;
  end if;
end;

$ cql --in a.sql --sem
Error at a.sql:7 : in call : CQL0344: argument must be a nullable type (but not constant NULL) in 'cql_inferred_notnull'
ENFORCE_STRICT NOT NULL AFTER CHECK;

CREATE PROC improvements_work_within_ctes ()
BEGIN
  DECLARE a INTEGER;
  IF a IS NOT NULL THEN
    WITH RECURSIVE
    foo (b) AS (SELECT cql_inferred_notnull(cql_inferred_notnull(a)))
    SELECT b
      FROM foo;
  END IF;
END;
semantic errors present; no code gen.
```

With this diff, we end up with `SELECT cql_inferred_notnull(a)`, as expected.

Reviewed By: ricomariani

Differential Revision: D29424108

fbshipit-source-id: d983d4bbc530a68064c747cb2c43dc1257de2e26