when upgrading schema, keep the facet state in a little side hash table so we don't have to query O(n) times
Summary:
The general idea here is to store the content of the facets table in a small side hash table so that we don't have to keep querying it.  There can be hundreds of such facets in a real schema and all that parsing and binding is excessive.

There was a light refactor of the code to allow for easy exception handling and also to just make some of the procedures a little smaller.  All the diff logic is together for instance.

Most of the procedures in the upgrader were marked private.  This saves us some space and prevents people from calling into the internals.

This is all intended to be a big no-op in the end.  Just saving some work.

Note the hashtable in the runtime has a very limited capability set and isn't really intended for general use and probably interoperates with nothing in your system.  Most users will not want to use this except for highly low level operations like in the schema upgrader but technically it is public.

Reviewed By: winniequinn

Differential Revision: D29480129

fbshipit-source-id: 9670adecc1b902df2c0f550699567e6a5c11e759