Account for ownership transfer when returning bridged Obj-C result set copy
Summary: Additional logic may be required when this ownership transfer occurs - for example, if the C result set is reference-counted, the count may need to be decremented.  A custom cqlrt now has the ability to note that it lost ownership of an object which is useful especially in test code.

Reviewed By: ricomariani

Differential Revision: D29359596

fbshipit-source-id: 075eaaf73a367f0760cbb1374b6959662cc7da29