Do not emit or parse cql_inferred_notnull
Summary: Since `cql_inferred_notnull` should only appear as the product of a rewrite rule and should never be in a source program, it didn't really make sense that that we were emitting it at all, and doing so was causing problems. Here, we both disallow its inclusion in programs and eliminate any visibility of the function to users.

Reviewed By: ricomariani

Differential Revision: D29482969

fbshipit-source-id: 073d1b12b3c3376960befbaf7ffcfd62c1c995da