CQL: fixing order of operations...
Summary:
You would think after nearly 4 years order of operations would be right... but no.....

Despite the fact that ast.h CLEARLY has experiments that show where NOT belongs we did not in fact put it there.
NOT is *weaker* than BETWEEN it has to go above there.  And likewise the token order in cql.y has to match
the order and again we didn't get this right.

If the priorities in the #define are wrong then the generated sql is wrong.

If the tokens bindings are wrong then the parsed ast is wrong...

I added a bunch more test for binding strength of not and between

I also made it so that all the existing run tests that verify order of operations also check the SQL version of the tree.

These run tests did in fact catch the problem with NOT and BETWEEN but it's all fixed now.

Reviewed By: toddkrabach

Differential Revision: D29698054

fbshipit-source-id: f6782bd2f239583cd1f1ebcfb2fea2277a480a83