Enable a bunch more expectations to run in both SQLite context and CQL context
Summary:
Just switching a few more expects to EXPECT_SQL_TOO this evaluates (x) and (select x) both of which should pass.

This is presently all passing but if we had done this some time ago we would have found bugs with the SQL generation so enabling these for future protection.

Reviewed By: toddkrabach

Differential Revision: D29715959

fbshipit-source-id: fd134368d14de9ba76ac0035b6213a13b724b66f