Further audit of order of ops for much more exotic cases
Summary:
I made a comment in the sqlite forums https://sqlite.org/forum/forumpost/daaad9e1ab suggesting some
additional documentation.  I got a pointer to the grammar for sqlite which has the definitive order:

```
%left OR.
%left AND.
%right NOT.
%left IS MATCH LIKE_KW BETWEEN IN ISNULL NOTNULL NE EQ.
%left GT LE LT GE.
%right ESCAPE.
%left BITAND BITOR LSHIFT RSHIFT.
%left PLUS MINUS.
%left STAR SLASH REM.
%left CONCAT.
%left COLLATE.
%right BITNOT.
```

This was so close to what I had... but not quite.  BETWEEN needed a tiny adjustment.

I also realized that I had to keep the NOT_LIKE, NOT_IN, and NOT_BETWEEN tokens to make sure that they got
the priority of equality ops and not the priority of "NOT" so I added extra tests for NOT BETWEEN and friends.

Reviewed By: RaoulFoaleng

Differential Revision: D29748437

fbshipit-source-id: 9acccfa92f43ee1ba6e390aafb586f6486cf3b4a