Use left recursion on statement lists to avoid stack overflow
Summary:
This is in response to https://github.com/facebookincubator/CG-SQL/issues/65

Thanks to mingodad for uncovering this issue and suggesting a remedy.

In the event that there are many (several thousand) statements in a procedure we could explode the bison stack because we use right recursion.

To avoid this problem, stmt_list is restated with left recursion.  This requires us to keep track of the tail pointer so we use a trick, re-using the parent pointer for the tail while we cons the tree.  The parent pointer at the root is otherwise useless until the statement list is inserted.

The final statement list is wired as usual and so the trick is invisible to all later stages.  But we never need much stack now as we always reduce the list as we go.

Reviewed By: winniequinn

Differential Revision: D29802524

fbshipit-source-id: 5f2238af38d636e81c66553ba57f87e0ac16f30e