Allow ifnull_crash and ifnull_throw to improve nullability
Summary:
`ifnull_crash(x)` and `ifnull_throw(x)` should improve the nullability of `x` (up to, at most, the end of the current statement list) because evaluation will not continue locally if `x` is NULL.

This must be done very carefully, however, as short-circuiting expressions complicate matters. In particular, any improvements made within the right side of an `AND` or an `OR` must not persist beyond of the right subexpression because the subexpression might not be evaluated. The same goes for subexpressions on the right side of an `IN` operator.

Reviewed By: ricomariani

Differential Revision: D29810612

fbshipit-source-id: c658f8996f0d2c5faf62500840df2a353e23a3c0