add alternate java codegen with interfaces
Summary:
If `--java_fragment_interface_mode` is specified, base fragments and extension fragments will codegen a series of dependent interfaces, and the assembly fragments will generate a class that implements the top-most interface.  This makes the generated java match the same dependency chain as the source CQL procs.

In order to maintain these references between the layers, it uses a similar option to the `--java_assembly_query_classname` parameter.  Any fragment codegen that extends/implements an interface (that is, extension and assembly fragments) may specify `--java_fragment_interface` to supply the interface that it depends on.  If none is specified, the codegen will assume the current namespace (of the fragment being generated) and use the previous proc name (the top-most base or extension fragment in the dependency chain).  If the proc name does not match the file name, it is required that the interface be supplied as a parameter, or the java will have an incorrect reference and will not be able to be built.

Reviewed By: ricomariani

Differential Revision: D29684665

fbshipit-source-id: fc5383fc8842a4ddfe48f925af9d9f49123690a8