update param parsing code for consistency and to provide more consistent error messages
Summary: The params that require N additional params have a helper function that gathers the remaining params.  This makes a similar helper for params that require 1 additional param, making the processing of them more consistent, and formatting the error message with the same prefix for each (still allows for a custom suffix per param, for better context).

Reviewed By: ricomariani

Differential Revision: D29720973

fbshipit-source-id: d161b22be9b4ac3189cc44dc1dd96cc5e1c9498d