Adding a run test to be sure that IS TRUE is working correctly
Summary:
https://sqlite.org/forum/forumpost/70e78ad16a

CQL takes the position that IS TRUE is an operator

SQLite's current position is that IS TRUE is not an operator but the IS operator gives
you IS TRUE semantics if the right operand is the boolean literal TRUE.

This is similar for IS FALSE, IS NOT TRUE, etc.

SQLite is getting this wrong so CQL conforms to what all the other SQL's do and corrects it in the output (which
happens for free anyway).

Reviewed By: winniequinn

Differential Revision: D29861165

fbshipit-source-id: cd8074e0bc2b32af2d5f85cc1304b6b53080d943