Support nonnull improvements for IIF
Summary: This has the side-effect of fixing a crash on the expression `IIF(a IS NOT NULL, ifnull_crash(a), 42)`. Previously, its arguments would type check just fine before the rewrite, but then fail to type check after the rewrite due to `ifnull_crash` being called on a then-improved `a` (which came into effect due to `CASE` performing the improvement), then we'd crash on `Invariant(!is_error(ast))` in `rewrite_iif`.

Reviewed By: ricomariani

Differential Revision: D29927011

fbshipit-source-id: 190986944b5b443540b69d3bdd9c3907a2d0fba9