Factor out handling of special functions
Summary:
Certain functions are special in that they require atypical semantic analysis of their arguments. We've been handling this with ad hoc checks in `sem_expr_call`, but as the number of special functions continues to grow, this will become increasingly ugly.

The rationale for doing this now is that `iif` is broken because its arguments are being analyzed by `sem_expr_call` absent any potential improvements from the conditional, then re-analyzed *with* potential improvements as part of the case rewrite. This can actually crash the compiler in some cases (as in the case of `IIF(a is not null, ifnull_crash(a), 42)`, where `a` will be of a `NOT NULL` type when `ifnull_crash(a)` is re-analyzed as part of the rewrite).

The above issue will be addressed in follow-up diff.

Reviewed By: ricomariani

Differential Revision: D29872135

fbshipit-source-id: a213fa9f57305d533467a8b5d8c249704535fa6f