Clean up special function analysis
Summary: The analysis functions for CQL functions declared with `SPECIAL_FUNC_INIT` now indicate that they are aggregate functions by setting `*is_aggregate` to true instead of returning a value of type `special_func_kind`. This reduces noise and ensures that the typically desired behavior of not treating special functions as aggregate functions is the default.

Reviewed By: ricomariani

Differential Revision: D29949769

fbshipit-source-id: ea8cb0f57c6c1aa30d40255af70a13a9a220630f