strict procedure names becomes the norm, you can't call undeclared functions at all now.
Summary:
It used to be that if you tried to call an undeclared function you would get a normal C style call.  This made it possible to call things like printf without any declarations.  That turns out to be very dangerous.

The new norm is that the old `enable_strict procedure` is now the only behavior (you must declare).  If you want to call an external C thing with varargs you have to do `declare proc printf no check`.  This gives you C style args (like c strings) in the call.  There's no change to the code gen it's just that you have to be explicit about when you want this now.  In general this should be avoided.

Note that CQL can't generate the C prototype for such a function, so normally you only call things that are in the set of things that are in your cqlrt.h.  You can use `fuzic` to add more includes if needed but this rapidly becomes a slipperly slope.  This feature should be avoided except for essential runtime functions with clear wrapping procs.

Reviewed By: winniequinn

Differential Revision: D29950965

fbshipit-source-id: 545bbf491ca12e7628b38aa2f73bb8d28962ef76