Tighten if_stmt AST assumptions
Summary: Some portions of the code were checking to see if an `if_alt` existed within an `if_stmt`, while others were asserting that the `if_alt` must always be present (which, indeed, it must). This simply makes the assumption universal and eliminates a couple of unneeded ifs. A handful of places in which a `cond_action` could have been extracted via `EXTRACT_NOTNULL` were changed as well.

Reviewed By: ricomariani

Differential Revision: D29953350

fbshipit-source-id: 26a914c054a657d04cca081ff9a8dfef241d0581