use message dedupe proxy in deserializer
Summary:
For now, we dedupe based off the persistent ID. As we move away from that, we will need to dedupe using transport IDs instead.

This sproc uses the dedupe proxy implemented in the previous diff to do that. This should cover any issues we might have with getting messages from encrypted backups where the persistent ID is not necessarily consistent.

Differential Revision: D30055454

fbshipit-source-id: 6086e358228ac1559009f5fcc57c9af97b9145e9