Generalize guard improvements
Summary:
Previously, we only performed nullability improvements after guards that were literal guard statements of the form `IF expr control_stmt`. This generalizes improvements to all IFs that follow the guard pattern; that is, IFs without ELSE IF or ELSE branches where the THEN branch ends with a control statement. For example:

```
declare a int;
if a is null then
  call printf("Hello, world!\n");
  return;
end if;
-- `a` has type INT NOT NULL here
```

Reviewed By: ricomariani

Differential Revision: D29955045

fbshipit-source-id: 8bb90487be0148bfb1737fca99bfec921de28cfc