Adjust railroad diagram tweaks
Summary:
The grammar text was recently modified for clarity in the doc but that makes the railroad diagram more verbose.

These grammar tweaks covert the repeated NOT IN, NOT LIKE, etc nodes to optional NOT syntax for just the railroad diagram.  This was done before but the patterns needed to be adjusted for the new form.  The railroad diagram is back to looking like it used to.

Note the current diagram is not wrong, it's just more verbose than it needs to be.

Reviewed By: RaoulFoaleng

Differential Revision: D30140225

fbshipit-source-id: 901d91626fb01dc862ba261338f1f56433fd472d