Emit a CRC for each kind of schema item into the JSON -- allows for change detection
Summary:
This CRC includes all aspects of the table/view/index/trigger/ad_hoc_migrator text so you can use it as a quick sanity check for sameness of items between deployers

Note: to avoid any possible data loss on 32 bit systems, the CRC is emitted as a string.  Typically exact equality is required anyway so it hardly matters.

Reviewed By: RaoulFoaleng

Differential Revision: D30148079

fbshipit-source-id: 183522aecb9ab6d0463577af0e856c5e20dae41d