Improve rendering of keywords in docs
Summary: We were showing tokens like `NULL_`  and `FOR_EACH_ROW` in the "Statement/Type Keywords" section of the documentation despite showing replaced versions like `"NULL"` and `"FOR EACH ROW"` in the "Rules" section of the documentation. This applies replacements to the "Statement/Type Keywords" section so that both sections are in alignment. It also alphabetically sorts the keywords.

Reviewed By: ricomariani

Differential Revision: D30236622

fbshipit-source-id: 02f6b227c7f56b99817f39425b8e466a5f1ab707