Allow false conditions to improve later branches in IF
Summary:
Due to earlier work, IFs introduces improvements for each branch with the knowledge that its associated condition must have been true if the branch's statements are being evaluated. This change allows IF to go further and improve a branch with the knowledge that, if it is being taken, all previous branches' conditions must have been false.

One benefit of this change is it is no longer necessary for a programmer to "flip" a conditional to get an improvement. The following two are now equivalent:

```
IF a IS NOT NULL THEN
  -- `a` is improved here
ELSE
  ...
END IF;
```

```
IF a IS NULL THEN
  ...
ELSE
  -- `a` is improved here
END IF;
```

Reviewed By: ricomariani

Differential Revision: D30383153

fbshipit-source-id: 6417d9d2ec8ba4b88ad70121371100c35585b9e0