Require Something better than Flex 2.5 due to yylineno bugs
Summary:
trailing context with newlines caused the line number to be off.  This is
causes catastrophic failure of the tests because they use the line numbers
to find the match patterns.  But also this causes bogus line number reports
to users if any trailing context had a newline in it.  Easy repro:

```
let x := 1 is true
;
```
all line numbers are now +1 because the \n after true will be processed twice.  Each time this happens the lines are an additional +1 diverged from reality.

This is fixed later versions of flex.

Reviewed By: RaoulFoaleng

Differential Revision: D30476059

fbshipit-source-id: 7a74d2339ed0eb94b12cee8b71440c83e25e17ab