Use trailing context and whitespace consistently in multi-word tokens
Summary:
This is mostly a no-op change, the syntax was changed to use the {abbreviation} form
in FLEX. However, a modern (2.6+) flex is required to do this change because the
"UNION ALL" token is routinely the last token on a line.  With the correct trailing context
the "\n" character was consistently read twice, throwing the lines off.

Flex 2.6 fixes this, this means that, as with bison, the xcode version of flex is too old.
I verified that "brew install flex" is good enough.

Reviewed By: RaoulFoaleng

Differential Revision: D30476074

fbshipit-source-id: b0ea7afba81d1d9eb0cf61acab3b16c8f8f35da1