Respect OUT args in conditions when improving
Summary:
CQL already removed nullability improvements for identifiers when they were used as OUT arguments. There was, however, a corner case: If an identifier was used as an OUT argument *within the same conditional* in which the opportunity for its improvement was discovered, *and* the OUT usage came *after* the expression triggering the improvement, the improvement would be made even though it was not safe to do so. For example, `a` cannot be safely improved in the following code, and this change ensures it no longer will be:

```
IF a IS NOT NULL AND requires_out_returns_bool(a) THEN
  -- `a` can't be improved here because the mutation occurs after the check
END IF;
```

On the other hand, this is perfectly safe and continues to work as expected:

```
IF requires_out_returns_bool(a) AND a IS NOT NULL THEN
  -- `a` is improved here because the check occurs after the mutation
END IF;
```

Reviewed By: ricomariani

Differential Revision: D30528637

fbshipit-source-id: 26de1549d19bd028e85d35107cf1c56461fa8148