Support branch-independent nullability improvements
Summary:
It was the case that un-improvements resulting from hazards (such as those due to setting a variable to null) within one branch of a set of branches would negatively impact all later branches. This change introduces contingent nullability contexts that are reverted upon the completion of their analyses such that branches are now effectively analyzed independently.

Support for CASE and IIF will be added in a follow-up.

Reviewed By: ricomariani

Differential Revision: D30589378

fbshipit-source-id: f8ac6cbede0cdfee952f7f4bdc81cedb988dc709