Disallow improvements in CASE expressions that match
Summary:
Recently, nullability improvements were added for CASE expressions. These were intended to work the same way they do for IF: Assuming each condition is true, improve the associated expression accordingly.

Unfortunately, nullability improvements were inadvertently possible in `CASE` expressions that matched upon a value. In such situations, a branch having been taken does not necessarily imply that its associated condition was true. In fact, in such matching CASE expressions, the condition is not a condition at all, but rather a value to be compared to some expression. Should that expression be equivalent to false, a branch would be taken when its associated expression to test was also false. For example:

```
CASE FALSE                  -- note the `FALSE` expression to match
  WHEN x IS NOT NULL THEN x -- shouldn't improve `x` as `x IS NOT NULL` is false
                            -- when this branch is taken
  ELSE ...
END
```

Given this, we were in a situation where improvements were being made unsafely whenever the expression to match was equivalent to false. To avoid this, we disallow improvements whenever the matching form of CASE is used.

Reviewed By: ricomariani

Differential Revision: D30659926

fbshipit-source-id: 6ba6cbb0b0b8bdeb6bad52ff5e230da103d7439b