Add test for left-to-right argument evaluation
Summary:
We depend upon left-to-right argument evaluation for nullability improvments to be safe. For example:

```
IF x IS NULL RETURN;
CALL some_proc(
  requires_notnull(x), -- must be evaluated first
  requires_out_arg(x)  -- must be evaluated second
);
```

We can only use `x` at a nonnull type in the call to `requires_notnull` if left-to-right evaluation is guaranteed. If it were not, `requires_out_arg(x)` could be evaluated first, and that could set `x` to `NULL` before `requires_notnull` was called.

This adds a new codegen test to ensure evaluation is left-to-right in this particular case.

Unfortunately, there exist other cases where evaluation is not left-to-right as required. These will be addressed in a follow-up.

Reviewed By: RaoulFoaleng

Differential Revision: D31102614

fbshipit-source-id: 785704c76435ffb5cb88894c271e41be0ca2de1e