Handle improvements for result columns properly
Summary:
We want to improve `y` in `SELECT x AS y FROM t WHERE y IS NOT NULL`, but only when `y` is not a column in `t`. This is because a `WHERE` clause sees the `FROM` clause first and the aliases in the expression list second; if `y` is a column in `t`, `y IS NOT NULL` refers to that column, not to the alias `y` of `x`.

The logic that existed before made a couple of incorrect assumptions about how binding rules worked in SQLite. Those problems are remedied here. See `sem_select_expr_list_con` for a deep explanation of the new approach.

As a bonus, this change also fixes a bug where `SELECT nullable(x) FROM t WHERE x IS NOT NULL` would result in a nonnull result column `x` (because any column named `x` in the result was being erroneously improved). One previously had to work around this by using an alias to prevent the improvement, e.g., `SELECT nullable(x) as x FROM t WHERE x IS NOT NULL`.

Reviewed By: ricomariani

Differential Revision: D31039640

fbshipit-source-id: 19e366ea50df3ada34e4ee3ce24b0664bccf98e1