Support nonnull improvements in tests and make it a no-op
Summary:
NOTE: This fixes a recently introduced issue in which CQL's tests would fail to run due to always-on nullability improvements.

The presence of `enforce_strict not null after check` was always intended to be temporary until nullability improvements became the default. Now that we're ready to make the transition, we can remove most traces of this option, convert CQL's tests to work with it always enabled, and leave only no-op parsing behind for the sake of (brief) partial backwards compatibility.

Reviewed By: ricomariani

Differential Revision: D30982099

fbshipit-source-id: 3bed61fcb9e96f3b189dd51ec50192ecb5bd642b